+++ src/main/java/com/upfpo/app/entity/ComplaintsComments.java
@@ -11,7 +11,7 @@ public class ComplaintsComments {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name="id")
-    private Long id;
+    private Integer id;
 
     private String comment;
 
@@ -23,18 +23,18 @@ public class ComplaintsComments {
     @JsonIgnore
     private Complaints complaints;
 
-    public ComplaintsComments(Long id, String comment, String createBy, Complaints complaints) {
+    public ComplaintsComments(Integer id, String comment, String createBy, Complaints complaints) {
         this.id = id;
         this.comment = comment;
         this.createBy = createBy;
         this.complaints = complaints;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
