+++ src/main/java/com/upfpo/app/repository/ComplaintRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComplaintRepository extends JpaRepository<Complaints, Long> {
+public interface ComplaintRepository extends JpaRepository<Complaints, Integer> {
 }
